Cleanup of comments and tests
