+++ tests/test_controller.py
@@ -3,7 +3,6 @@ import os
 import shutil
 import unittest
 from types import SimpleNamespace
-from unittest.case import skip
 from unittest.mock import ANY, MagicMock
 
 import numpy as np
@@ -701,47 +700,6 @@ class TestController(unittest.TestCase):
         self.assertEqual(2, actual.room_dimensions.y)
         self.assertEqual(3, actual.room_dimensions.z)
         self.assertEqual(None, actual.goal)
-
-    # DW: moved to sceneconfigurationschema @post_load
-    @skip
-    def test_remove_none(self):
-        actual = self.controller._remove_none({})
-        self.assertEqual({}, actual)
-
-        actual = self.controller._remove_none({"test": None})
-        self.assertEqual({}, actual)
-
-        actual = self.controller._remove_none({"test": 1})
-        self.assertEqual({"test": 1}, actual)
-
-        actual = self.controller._remove_none({"test": 1, "none": None})
-        self.assertEqual({"test": 1}, actual)
-
-        actual = self.controller._remove_none({"test1": {"test2": 1}})
-        self.assertEqual({"test1": {"test2": 1}}, actual)
-
-        actual = self.controller._remove_none(
-            {"test1": {"test2": 1, "none": None}})
-        self.assertEqual({"test1": {"test2": 1}}, actual)
-
-        actual = self.controller._remove_none(
-            {"test1": {"test2": 1}, "none": None})
-        self.assertEqual({"test1": {"test2": 1}}, actual)
-
-        actual = self.controller._remove_none(
-            {"test1": {"test2": 1, "none": None}, "none": None})
-        self.assertEqual({"test1": {"test2": 1}}, actual)
-
-        actual = self.controller._remove_none({"test1": [{"test2": None}]})
-        self.assertEqual({"test1": [{}]}, actual)
-
-        actual = self.controller._remove_none(
-            {"test1": [{"test2": None, "test3": "test"}]})
-        self.assertEqual({"test1": [{"test3": "test"}]}, actual)
-
-        actual = self.controller._remove_none(
-            {"test1": [{"test2": None}], "test3": False})
-        self.assertEqual({"test1": [{}], "test3": False}, actual)
 
     def test_set_config(self):
         first_config = self.controller._config
